+++ src/charm.py
@@ -29,8 +29,9 @@ from ops.model import (
     WaitingStatus,
 )
 
+from resource_manager.decorator import InsufficientPermissionError
 from resource_manager.gateway import CreateGatewayError, GatewayResourceManager
-from resource_manager.resource_manager import InsufficientPermissionError, InvalidResourceError
+from resource_manager.resource_manager import InvalidResourceError
 from state.config import CharmConfig, InvalidCharmConfigError
 from state.gateway import GatewayResourceDefinition
 from state.tls import TLSInformation, TlsIntegrationMissingError
