+++ src/state/validation.py
@@ -12,6 +12,7 @@ import ops
 import state
 import state.config
 import state.tls
+from resource_manager.decorator import InsufficientPermissionError
 
 logger = logging.getLogger(__name__)
 
@@ -61,12 +62,13 @@ def validate_config_and_integration(
                 state.config.InvalidCharmConfigError,
                 state.tls.TlsIntegrationMissingError,
                 state.config.GatewayClassUnavailableError,
+                InsufficientPermissionError,
             ) as exc:
                 if defer:
                     event: ops.EventBase
                     event, *_ = args
                     event.defer()
-                logger.exception("Wrong Charm Configuration")
+                logger.exception(str(exc))
                 instance.unit.status = ops.BlockedStatus(str(exc))
                 return None
